+++ src/ape_accounts/accounts.py
@@ -142,7 +142,10 @@ class KeyfileAccount(AccountAPI):
               If not provided, you will be prompted to enter it.
         """
         self.unlock(passphrase=passphrase)
-        logger.warning("Danger! This account will now sign any transaction its given.")
+
+        if enabled:
+            logger.warning("Danger! This account will now sign any transaction its given.")
+
         self.__autosign = enabled
 
     def _prompt_for_passphrase(self, message: Optional[str] = None, **kwargs):
